+++ weldx/constants.py
@@ -2,8 +2,9 @@
 from pathlib import Path as _Path
 
 import pint
+import pint_xarray  # noqa: F401 # pylint: disable=W0611
 
-UNIT_KEY = "units"  # default nomenclature for storing physical units information
+UNITS_KEY = "units"  # default nomenclature for storing physical units information
 
 WELDX_PATH = _Path(__file__).parent.resolve()
 
@@ -63,11 +64,14 @@ U_.__doc__ = """For details on working with quantities and units, please see the
 """
 
 
+# set default unit registry for pint-xarray
+pint.set_application_registry(WELDX_UNIT_REGISTRY)
+
 __all__ = (
     "WELDX_PATH",
     "WELDX_QUANTITY",
     "WELDX_UNIT_REGISTRY",
     "Q_",
     "U_",
-    "UNIT_KEY",
+    "UNITS_KEY",
 )