+++ CHANGELOG.rst
@@ -9,6 +9,8 @@
 added
 =====
 
+-  add experimental unit support for ``.weldx.interp_like`` accessor [:pull:`518`]
+
 removed
 =======
 
@@ -33,6 +35,8 @@ deprecations
 
 dependencies
 ============
+
+-  add ``pint-xarray`` dependency [:pull:`518`]
 
 ********************
  0.5.2 (18.11.2021)